made performance benchmarks more closely match dappers
#10 excellent first steps

this is a good step towards the coveted mark of sql2os time of ~75ms.
the bulk of the 250ish ms spiffy consumes in the benchmark is the
initial time to prime the caches with reflection data. There's gotta be
a better way to do this, and I'll get there.